Move c2client errors to errors.py
